script: Fix install-flynn behaviour in presence of whitespace
There were a handful of places in the install-flynn script that would
do the wrong thing if given a string with a space in it. Mostly,
this would manifest as a series of args like "foo bar", "baz" being
interpreted as "foo", "bar", "baz".
I don't believe there were any legitimate use cases for strings containing
whitespace, but this change will improve error messages if that does occur.
In some cases the strings in question were hard-coded and do not contain
whitespace, but these changes still improve robustness and style.

Signed-off-by: Mike Lang <mikelang3000@gmail.com>